Add scrollpane to wrap overflow frames, send trasmission not possible jpeg when images too big to send over udp

+++ client/src/main/java/pl/nieruchalski/client/domain/service/ConnectionService.java
@@ -57,7 +57,7 @@ public class ConnectionService {
                 throw new HostRefusedAccessCodeException();
             }
             Integer clientUdpPort = inputStream.readUnsignedShort();
-            outputStream.writeInt(this.udpSocket.getLocalPort());
+            outputStream.writeShort(this.udpSocket.getLocalPort());
             ViewerHost host = new ViewerHost(socket, clientUdpPort);
             HostManager.getInstance().registerHost(host);
             NewConnectionPublisher.getInstance().broadcast(host);
